+++ euphonic/cli/intensity_map.py
@@ -4,7 +4,7 @@ from typing import Optional
 import matplotlib.style
 
 import euphonic
-from euphonic import ForceConstants, QpointFrequencies, ureg
+from euphonic import ForceConstants, QpointPhononModes, ureg
 import euphonic.plot
 from euphonic.styles import base_style
 from euphonic.util import get_qpoint_labels
@@ -31,14 +31,18 @@ def main(params: Optional[list[str]] = None) -> None:
     frequencies_only = (args.weighting != 'coherent')
     data = load_data_from_file(args.filename, verbose=True,
                                frequencies_only=frequencies_only)
-    if not frequencies_only and isinstance(data, QpointFrequencies):
+    if (not frequencies_only
+        and not isinstance(data, (QpointPhononModes, ForceConstants))
+    ):
         raise TypeError('Eigenvectors are required to use '
                         '"--weighting coherent" option')
-    if args.weighting.lower() == 'coherent' and args.temperature is not None:
-        if not isinstance(data, ForceConstants):
-            raise TypeError('Force constants data is required to generate '
-                            'the Debye-Waller factor. Leave "--temperature" '
-                            'unset if plotting precalculated phonon modes.')
+    if (args.weighting.lower() == 'coherent'
+        and args.temperature is not None
+        and not isinstance(data, ForceConstants)
+    ):
+        raise TypeError('Force constants data is required to generate '
+                        'the Debye-Waller factor. Leave "--temperature" '
+                        'unset if plotting precalculated phonon modes.')
 
     q_spacing = _get_q_distance(args.length_unit, args.q_spacing)
     recip_length_unit = q_spacing.units