+++ tests_and_analysis/test/euphonic_test/test_force_constants.py
@@ -44,7 +44,7 @@ class ExpectedForceConstants:
 
     @property
     def born(self):
-        if 'born' in self.data.keys():
+        if 'born' in self.data:
             return np.array(self.data['born'])*ureg(
                 self.data['born_unit'])
         return None
@@ -56,7 +56,7 @@ class ExpectedForceConstants:
 
     @property
     def dielectric(self):
-        if 'dielectric' in self.data.keys():
+        if 'dielectric' in self.data:
             return np.array(self.data['dielectric'])*ureg(
                 self.data['dielectric_unit'])
         return None
@@ -392,8 +392,9 @@ class TestForceConstantsCreation:
         real_import = builtins.__import__
         def mocked_import(
                 name, globals, locals, fromlist, level):  # noqa: A002
-            if name == 'yaml':
-                if fromlist is not None and fromlist[0] == 'CSafeLoader':
+            if (name == 'yaml'
+                and fromlist is not None
+                and fromlist[0] == 'CSafeLoader'):
                     raise ImportError
             return real_import(name, globals, locals, fromlist, level)
         mocker.patch('builtins.__import__', side_effect=mocked_import)