+++ tests_and_analysis/test/euphonic_test/test_force_constants_calculate_qpoint_frequencies.py
@@ -79,7 +79,7 @@ class TestForceConstantsCalculateQPointFrequencies:
         # Only give gamma-acoustic modes special treatment if the acoustic
         # sum rule has been applied
         tol_kwargs = {}
-        if 'asr' in func_kwargs.keys():
+        if 'asr' in func_kwargs:
             tol_kwargs['acoustic_gamma_atol'] = 0.55
         # Use larger tolerances with reciprocal ASR - formalism works
         # only at gamma but is applied to all q, so problem is less