+++ tests_and_analysis/test/euphonic_test/test_qpoint_frequencies.py
@@ -343,8 +343,9 @@ class TestQpointFrequenciesCreation:
 
         def mocked_import(
                 name, globals, locals, fromlist, level):  # noqa: A002
-            if name == 'yaml':
-                if fromlist is not None and fromlist[0] == 'CSafeLoader':
+            if (name == 'yaml'
+                and fromlist is not None
+                and fromlist[0] == 'CSafeLoader'):
                     raise ImportError
             return real_import(name, globals, locals, fromlist, level)
         mocker.patch('builtins.__import__', side_effect=mocked_import)
@@ -632,7 +633,7 @@ class TestQpointFrequenciesCalculateDosMap:
                 qpt_freqs.frequencies[qpt].reshape(1, -1),
                 np.array([qpt_freqs.weights[qpt]]))
             dos_kwargs_single_qpt = copy.copy(dos_kwargs)
-            if 'mode_widths' in dos_kwargs.keys():
+            if 'mode_widths' in dos_kwargs:
                 dos_kwargs_single_qpt['mode_widths'] = dos_kwargs[
                     'mode_widths'][qpt].reshape(1, -1)
             dos_single_qpt = qpt_freqs_single_qpt.calculate_dos(