+++ tests_and_analysis/test/euphonic_test/test_spectrum1dcollection.py
@@ -38,15 +38,12 @@ class ExpectedSpectrum1DCollection:
 
     @property
     def x_tick_labels(self):
-        if 'x_tick_labels' in self.data.keys():
-            return [tuple(x) for x in self.data['x_tick_labels']]
-        return None
+            return (list(map(tuple, self.data['x_tick_labels']))
+                    if 'x_tick_labels' in self.data else None)
 
     @property
     def metadata(self):
-        if 'metadata' in self.data.keys():
-            return self.data['metadata']
-        return {}
+        return self.data.get('metadata', {})
 
     def to_dict(self):
         d = {'x_data': self.x_data.magnitude,