+++ tests_and_analysis/test/script_tests/test_intensity_map.py
@@ -1,3 +1,4 @@
+from contextlib import suppress
 import json
 import os
 from unittest.mock import patch
@@ -16,12 +17,10 @@ from tests_and_analysis.test.utils import get_castep_path, get_data_path
 pytestmark = pytest.mark.matplotlib
 # Allow tests with matplotlib marker to be collected and
 # deselected if Matplotlib is not installed
-try:
+with suppress(ModuleNotFoundError):
     import matplotlib.pyplot  # noqa: ICN001
 
     import euphonic.cli.intensity_map
-except ModuleNotFoundError:
-    pass
 
 quartz_json_file = get_data_path(
     'qpoint_phonon_modes', 'quartz',
@@ -133,7 +132,7 @@ class TestRegression:
         [quartz_no_evec_json_file, '--weighting', 'coherent']])
     def test_qpoint_frequencies_incompatible_args_raises_type_error(
             self, intensity_map_args):
-        with pytest.raises(TypeError):
+        with pytest.raises(TypeError, match="Eigenvectors are required"):
             euphonic.cli.intensity_map.main(intensity_map_args)
 
     @pytest.mark.parametrize('intensity_map_args', [
@@ -141,7 +140,7 @@ class TestRegression:
          '--temperature', '300']])
     def test_qpoint_modes_debyewaller_raises_type_error(
             self, intensity_map_args):
-        with pytest.raises(TypeError):
+        with pytest.raises(TypeError, match="Force constants data"):
             euphonic.cli.intensity_map.main(intensity_map_args)
 
     @pytest.mark.parametrize('intensity_map_args', [